% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{annotate_features}
\alias{annotate_features}
\title{Perform Annotation of Raw Data (Step 6)}
\usage{
annotate_features(raw_data)
}
\arguments{
\item{raw_data}{Raw data from proteome discoverer that has the Sample and Control
files indicated}
}
\value{
A modified data frame with appropriate modification to the raw data
}
\description{
Perform Annotation of Raw Data (Step 6)
}
\examples{
newdf<- annotate_features(raw_data)
}
