% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_calculations_pep.R
\name{area_calculations_pep}
\alias{area_calculations_pep}
\title{area_calculations_pep}
\usage{
area_calculations_pep(df_in)
}
\arguments{
\item{df_in}{A data frame with the following columns: MasterProteinAccessions,
Sequence, SampleControl, MOD, Precursor Abundance}
}
\value{
a data frame containing the values for the extent of modification
calculation
}
\description{
area_calculations_pep
}
\examples{
EOM<- area_calculations_pep(pd_data_annotated)
}
