% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_calculations_res.R
\name{area_calculations_res}
\alias{area_calculations_res}
\title{area_calculations_res}
\usage{
area_calculations_res(df_in)
}
\arguments{
\item{df_in}{data frame from modified raw data to calculate the EOM}
}
\value{
a data frame containing the extent of modification calculations
at the residue level
}
\description{
area_calculations_res
}
\examples{
EOM<- area_calculations_res(pd_data)
}
