% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/before_beginning.R
\name{before_beginning}
\alias{before_beginning}
\title{Before Analyzing Your Data be Sure to Run this Function}
\usage{
before_beginning()
}
\value{
Required  Packaged and Conflict Preferences
}
\description{
Before Analyzing Your Data be Sure to Run this Function
}
\examples{
before_beginning()
}
