% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_selectionLFQ.R
\name{column_selectionLFQ}
\alias{column_selectionLFQ}
\title{Column Selection for LFQ Data (Non TMT Labeled)}
\usage{
column_selectionLFQ(df)
}
\arguments{
\item{df}{}
}
\value{
A data frame with the selected columns.
}
\description{
Column Selection for LFQ Data (Non TMT Labeled)
}
\examples{
Selected_data <- column_selectionLFQ(raw_data)
}
