% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_selectionTMT.R
\name{column_selectionTMT}
\alias{column_selectionTMT}
\title{Select the Columns Necessary for Extent of Modification Calculation from TMT Data}
\usage{
column_selectionTMT(df)
}
\arguments{
\item{df}{}
}
\value{
A data frame with selected columns from the input TMT data frame
}
\description{
Select the Columns Necessary for Extent of Modification Calculation from TMT Data
}
\examples{
selected_data <- column_selectionTMT(TMT_data)
}
