% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_peptides_per_protein.R
\name{count_peptides_per_protein}
\alias{count_peptides_per_protein}
\title{Count the Number of Modified Peptides Per Protein}
\usage{
count_peptides_per_protein()
}
\value{
A Table and Graph of the Number of Modified Peptides Per Protein
}
\description{
Count the Number of Modified Peptides Per Protein
}
\examples{
count_peptides_per_protein()
}
