% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_residue_entries_per_protein.R
\name{count_residue_entries_per_protein}
\alias{count_residue_entries_per_protein}
\title{Count the Modified Residues Per Protein}
\usage{
count_residue_entries_per_protein()
}
\value{
An excel file of modified residues per protein and a bar graph of the data
}
\description{
Count the Modified Residues Per Protein
}
\examples{
count_residue_entries_per_protein()
}
