% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_graphing_df_pep.R
\name{filtered_graphing_df_pep}
\alias{filtered_graphing_df_pep}
\title{Filter Data for Creating Extent of Modification Graphs at the Peptide Level (Step 12)}
\usage{
filtered_graphing_df_pep(df_in)
}
\arguments{
\item{df_in}{A data frame with EOM data to filter for quantifiable
modifications}
}
\value{
a data frame containing the filtered FPOP data
}
\description{
Filter Data for Creating Extent of Modification Graphs at the Peptide Level (Step 12)
}
\examples{
graphing_data<- filtered_graphing_df_pep(EOM)
}
