% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPOP_Calculations.R
\name{FPOP_Calculations}
\alias{FPOP_Calculations}
\title{Calculate the Extent of Modification at the Peptide and Residue Level}
\usage{
FPOP_Calculations()
}
\value{
A list of dataframes that contain the Extent of Modification (EOM) and SD (Variance) at the peptide and residue level
}
\description{
Calculate the Extent of Modification at the Peptide and Residue Level
}
\examples{
FPOP_Calculations()
}
