% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eom_plot_pep.R
\name{generate_eom_plot_pep}
\alias{generate_eom_plot_pep}
\title{generate_eom_plot_pep}
\usage{
generate_eom_plot_pep(df_in, file_output, excel_filename)
}
\arguments{
\item{df_in}{a data frame with quantifiable peptide modifications}

\item{file_output}{The file path to save the graphs}

\item{excel_filename}{The file name to save the excel file}
}
\value{
Bar graphs corresponding to each quantifiable peptide in the input
data frame. Must Run save_graphs_pep() to save the graphs in the desired
location.
}
\description{
generate_eom_plot_pep
}
\examples{
plot_pep <- generate_eom_plot_pep(graphing_data = df_in,
file_output = file_output, excel_filename = excel_filename)
}
