% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eom_plot_res.R
\name{generate_eom_plot_res}
\alias{generate_eom_plot_res}
\title{Create a Plot for the Extent of Modification at the Residue Level (Step 16D)}
\usage{
generate_eom_plot_res(df_in, file_output, filename)
}
\arguments{
\item{df_in}{a data frame containing the EOM and SD values for each residue}

\item{file_output}{the directory where the output files will be saved}

\item{filename}{User defined file name}
}
\value{
a bar graph for each protein with quantifiable residue level
 modifications
}
\description{
Create a Plot for the Extent of Modification at the Residue Level (Step 16D)
}
\examples{
 generate_eom_plot_res(df_in = quant_graph_df_res, file_output = file_output,)
}
