% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_seq_metadata_res.R
\name{grab_seq_metadata_res}
\alias{grab_seq_metadata_res}
\title{Grab Sequence Metadata for Residue Level Graphing (Step 14)}
\usage{
grab_seq_metadata_res(df_in)
}
\arguments{
\item{df_in}{data frame with original data to match proteins and residue}
}
\value{
A data frame with columns MasterProteinAccessions, Sequence, Res,
start columns. ONLY used inside of graphing_data_res function.
}
\description{
Grab Sequence Metadata for Residue Level Graphing (Step 14)
}
\examples{
new_df<-grab_seq_metadata_res(raw_data)
}
