% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data from an Excel file (step 3 )}
\usage{
import_data()
}
\value{
A data frame with the desired data set
}
\description{
Import data from an Excel file (step 3 )
}
\examples{
import <- import_data()
}
