% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFQ_Prep.R
\name{LFQ_Prep}
\alias{LFQ_Prep}
\title{Prepare the Input Data for Calculating the Extent of Modification by
Selecting and Renaming the Relevant Columns and IDentifying the Start and End
of the Peptide Sequence}
\usage{
LFQ_Prep()
}
\value{
Data Frames Containing Annotated Data
}
\description{
Prepare the Input Data for Calculating the Extent of Modification by
Selecting and Renaming the Relevant Columns and IDentifying the Start and End
of the Peptide Sequence
}
\examples{
LFQ_Prep()
}
