% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_startend_res.R
\name{locate_startend_res}
\alias{locate_startend_res}
\title{locate_startend_res}
\usage{
locate_startend_res(raw_data)
}
\arguments{
\item{raw_data}{data frame with original data to map peptide locations on
protein sequences}
}
\value{
A data frame with with the start and ending peptide location
}
\description{
locate_startend_res
}
\examples{
new_df <- locate_startend_res(raw_data)
}
