% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_metedata_pep.R
\name{merge_metadata_pep}
\alias{merge_metadata_pep}
\title{Merge Metadata with Peptide Level Graphing Data (Step 11)}
\usage{
merge_metadata_pep(df_in, rawdatafastamerged)
}
\arguments{
\item{df_in}{Data frame with peptide level extent of modification calculations.}

\item{rawdatafastamerged}{Data frame with raw data and the fasta data merged.}
}
\value{
Data frame with peptide level extent of modification calculations and
metadata from the fasta file.
}
\description{
Merge Metadata with Peptide Level Graphing Data (Step 11)
}
\examples{
graphing_df_pep<- merge_metadata_pep(graphing_data_pep, rawdatafastamerged)
}
