% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_folder.R
\name{output_folder}
\alias{output_folder}
\title{Select the Output Directory for the Data (Step2)}
\usage{
output_folder()
}
\arguments{
\item{file_output}{The file path to save the output}
}
\value{
The path to the folder where the output will be saved.
}
\description{
Select the Output Directory for the Data (Step2)
}
\examples{
output_folder()
}
