\name{process_data}
\alias{file_path}
\title{select the file location of sequence searched MS sata}
\description{
  This function prompts the user to define the file path where an Excel file containing sequence searched MS data is located, reads the data, and performs necessary processing steps.
}
\usage{
  process_data()
}
\details{
  The function \code{process_data} does not take any arguments. Upon execution, it prompts the user to select an Excel file containing sequence searched MS data using a file dialog. It then reads the data from the chosen file using \code{read.xlsx} and performs any additional processing steps required.
}
\examples{
  # Call the function to process sequence searched MS data
  processed_data <- process_data()
  file<- process_data()
}
\note{
  Ensure that the 'xlsx' package is installed to read Excel files using 'read.xlsx'.
}
