% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dup.R
\name{remove_dup}
\alias{remove_dup}
\title{Remove Duplicate Sequest Node Identifications (Step 5)}
\usage{
remove_dup(df)
}
\arguments{
\item{df}{data frame with original data from PD to remove duplicates
created from using the multi-level sequence searching algorithm}
}
\value{
a data frame with no duplicate sequest node identifications
}
\description{
Remove Duplicate Sequest Node Identifications (Step 5)
}
\examples{
pd_data <- remove_dup(pd_data)
}
