% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_and_split_spectrum_files.R
\name{rename_and_split_spectrum_files}
\alias{rename_and_split_spectrum_files}
\title{Rename and Extract Spectrum File Identifiers}
\usage{
rename_and_split_spectrum_files(df_in)
}
\arguments{
\item{df_in}{}
}
\value{
A data frame with the renamed and split spectrum files.
}
\description{
Rename and Extract Spectrum File Identifiers
}
\examples{
modified_data <- rename_and_split_spectrum_files(Selected_data)
}
