% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns_interactively.R
\name{rename_columns_interactively}
\alias{rename_columns_interactively}
\title{Interactively Rename Columns Corresponding to TMT Data}
\usage{
rename_columns_interactively(data)
}
\arguments{
\item{data}{}
}
\value{
Data frame with the columns renamed by the user
}
\description{
Interactively Rename Columns Corresponding to TMT Data
}
\examples{
renamed_data <- rename_columns_interactively(selected_data)
}
