\name{resolve_conflicts}
\alias{resolve_conflicts}
\title{Resolve Package Conflict Preferences}
\description{
  This function resolves conflicts between package preferences for specific functions.
}
\usage{
  resolve_conflicts()
}
\arguments{
  \item{conflict_list}{A character vector containing the names of functions with conflicts.}
}
\value{
  A list showing resolved conflicts between packages for specified functions.
}
\details{
  The function \code{resolve_conflicts} checks a list of functions against predefined package preferences
  and returns the preferred package for each function if available, otherwise marks as "No preference".
}
\examples{
  # Define package preferences
  preferred_packages <- list(
    arrange = "dplyr",
    between = "dplyr",
    # ... (other preferences)
    read.fasta = "phylotools"
  )

  # Define the resolving function
  resolve <- function(conflict_list) {
    resolved_conflicts <- list()

    for (conflict in conflict_list) {
      if (conflict %in% names(preferred_packages)) {
        resolved_conflicts[[conflict]] <- preferred_packages[[conflict]]
      } else {
        resolved_conflicts[[conflict]] <- "No preference"
      }
    }

    return(resolved_conflicts)
  }

  # Final function to be used
  resolve_conflicts <- function() {
    return(resolve)
  }

  # Using the function
  resolve_conflicts()(c("arrange", "count", "failwith", "hour", "transpose", "read.fasta"))
}
\note{
  This function requires pre-defined preferences for function-package associations.
}
