% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleControl.R
\name{SampleControl}
\alias{SampleControl}
\title{Identify Sample (Laser) or Control (No Laser) Files (Step4)}
\usage{
SampleControl(pd_data)
}
\arguments{
\item{pd_data}{Sequence searched data to which a column can be added to indicate
Sample vs Control files by referencing the name indicated in the spectrum file.
Ensure that laser irradiated samples are indicated by "L" in the spectrum file.
Meanwhile control samples are indicated by "NL" in the spectrum file.}
}
\value{
A column indicating Sample vs Control
}
\description{
Identify Sample (Laser) or Control (No Laser) Files (Step4)
}
\examples{
pd_data <- SampleControl(pd_data)
}
