% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Prepare R by Installing the Necessary Packages; Eliminating Conflicts, Importing Data, and
Selecting an Output Folder}
\usage{
setup()
}
\value{
The imported FASTA file, raw data, and the selected output folder
}
\description{
Prepare R by Installing the Necessary Packages; Eliminating Conflicts, Importing Data, and
Selecting an Output Folder
}
\examples{
setup()
}
