% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_and_calculate_abundancesTMT.R
\name{sum_and_calculate_abundancesTMT}
\alias{sum_and_calculate_abundancesTMT}
\title{Calculate the Sum of Abundances and True Abundances for TMT Data}
\usage{
sum_and_calculate_abundancesTMT(data)
}
\arguments{
\item{data}{}
}
\value{
A data frame with the sum of abundances and true abundances for TMT data
}
\description{
Calculate the Sum of Abundances and True Abundances for TMT Data
}
\examples{
TMT_data <- sum_and_calculate_abundancesTMT(raw_data)
}
