% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables_and_Graphs.R
\name{Tables_and_Graphs}
\alias{Tables_and_Graphs}
\title{Generate and Save the Tables and Graphs for the LFQ Data}
\usage{
Tables_and_Graphs()
}
\value{
Tables as excel files and graphs as .PNG files corresponding to the LFQ data.
}
\description{
Generate and Save the Tables and Graphs for the LFQ Data
}
\examples{
Tables_and_Graphs()
}
