% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT_Prep.R
\name{TMT_Prep}
\alias{TMT_Prep}
\title{Prepare the Input TMT Data for Calculating the Extent of Modification by
Identifying the Sample vs Control Files and  Start and End of the Peptide Sequence}
\usage{
TMT_Prep()
}
\value{
Data Frames Containing Annotated TMT Data
}
\description{
Prepare the Input TMT Data for Calculating the Extent of Modification by
Identifying the Sample vs Control Files and  Start and End of the Peptide Sequence
}
\examples{
TMT_Prep()
}
