% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT_Quant.R
\name{TMT_Quant}
\alias{TMT_Quant}
\title{Prepare the Input Data TMT for Annotation and Quantification}
\usage{
TMT_Quant()
}
\value{
Data Frames Containing Quantified, Renamed, and Transformed TMT Columns
}
\description{
Prepare the Input Data TMT for Annotation and Quantification
}
\examples{
TMT_Quant()
}
