% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Transform the Horizontal TMT Data to a Linear Format}
\usage{
transform_data(data)
}
\arguments{
\item{data}{}
}
\value{
A data frame that has transformed the input data to a linear format
}
\description{
Transform the Horizontal TMT Data to a Linear Format
}
\examples{
transformed_data <- transform_data(renamed_data)
}
