% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_features.R
\name{annotate_features}
\alias{annotate_features}
\title{annotate_features}
\usage{
annotate_features(raw_data)
}
\arguments{
\item{raw_data}{raw data from proteome discoverer with modifications that
annotate specific features}
}
\value{
a modified data frame with appropriate modification to the raw data
}
\description{
annotate_features
}
\examples{
newdf<- annotate_features(raw_data)
}
