% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_calculations_pep.R
\name{area_calculations_pep}
\alias{area_calculations_pep}
\title{Calculate the Extent of Modification at the Peptide Level (Step 9)}
\usage{
area_calculations_pep(df_in)
}
\arguments{
\item{df_in}{The annotated data frame merged with the parsed FASTA file.}
}
\value{
a data frame containing the values for the extent of modification
calculation
}
\description{
Calculate the Extent of Modification at the Peptide Level (Step 9)
}
\examples{
EOM<- area_calculations_pep(pd_data_annotated)
}
