\name{area_calculations_res}
\alias{area_calculations_res}
\title{Calculate Area and Associated Metrics}
\description{
  This function performs calculations on a data frame, specifically focusing on area calculations and associated metrics.
}
\usage{
  area_calculations_res(df_in)
}
\arguments{
  \item{df_in}{A data frame containing the input data.}
}
\value{
  Returns a modified data frame with area calculations and associated metrics.
}
\examples{
  # Call the function with a data frame pd_data_fasta_merged
  Areas_res <- area_calculations_res(pd_data_fasta_merged)
}
