% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_totals_tablelist.R
\name{create_totals_tablelist}
\alias{create_totals_tablelist}
\title{Creating a List of Total Peptides and Residues Modified (Step 17)}
\usage{
create_totals_tablelist(df_in, df_res)
}
\arguments{
\item{df_in}{Filtered peptide level data frame}

\item{df_res}{Filtered residue level data frame}
}
\value{
A data frame containing the total number of unique proteins, sequences, and residues modified
}
\description{
Creating a List of Total Peptides and Residues Modified (Step 17)
}
\examples{
Totals<- create_totals_tablelist(quant_graph_df, quant_graph_df_re)
}
