% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTA_file.R
\name{FASTA_file}
\alias{FASTA_file}
\title{FASTA_file}
\usage{
FASTA_file()
}
\arguments{
\item{FASTA_path}{FASTA file path}
}
\value{
FASTA file saved as a data frame
}
\description{
FASTA_file
}
\examples{
HumanFASTA <- FASTA_file(file.fasta)
}
