% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_graphing_df_pep.R
\name{filtered_graphing_df_pep}
\alias{filtered_graphing_df_pep}
\title{filtered_graphing_df_pep}
\usage{
filtered_graphing_df_pep(df_in)
}
\arguments{
\item{df_in}{a data frame with EOM data to filter for quantifiable
modifications}
}
\value{
a data frame containing the filtered FPOP data
}
\description{
filtered_graphing_df_pep
}
\examples{
graphing_data<- filtered_graphing_df_pep(EOM)
}
