% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtered_graphing_df_res.R
\name{filtered_graphing_df_res}
\alias{filtered_graphing_df_res}
\title{filtered_graphing_df_res}
\usage{
filtered_graphing_df_res(df_in)
}
\arguments{
\item{df_in}{a data frame with EOM data to filter for quantifiable residue
level modifications}
}
\value{
A data frame containing the filtered data
}
\description{
filtered_graphing_df_res
}
\examples{
graphing_df_res <- filtered_graphing_df_res(graphing_df)
}
