% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eom_plot_res.R
\name{generate_eom_plot_res}
\alias{generate_eom_plot_res}
\title{generate_eom_plot_res}
\usage{
generate_eom_plot_res(df_in, file_output, excel_filename)
}
\arguments{
\item{df_in}{a data frame containing the EOM and SD values for each residue}

\item{file_output}{the directory where the output files will be saved}

\item{excel_filename}{the name of the excel file that the data was read from}
}
\value{
a bar graph for each protein with quantifiable residue level
 modifications
}
\description{
generate_eom_plot_res
}
\examples{
 generate_eom_plot_res(df_in, file_output, excel_filename)
}
