% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_bar_plot_pep.R
\name{generate_grouped_bar_plot_pep}
\alias{generate_grouped_bar_plot_pep}
\title{Create a Grouped Bar Graph for Peptide Level Analysis of Different Conditions}
\usage{
generate_grouped_bar_plot_pep()
}
\arguments{
\item{df_in}{The data frame containing the peptide level analysis data}

\item{file_output}{The user defined directory to save the grouped bar graphs}

\item{filename}{The user defined file name to use for the grouped bar graphs}
}
\value{
A grouped bar graph for each protein in the data frame
}
\description{
Create a Grouped Bar Graph for Peptide Level Analysis of Different Conditions
}
\examples{
generate_grouped_bar_plot_pep(df_in, file_output, filename)
}
