% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_seq_metadata_pep.R
\name{grab_seq_metadata_pep}
\alias{grab_seq_metadata_pep}
\title{Grab Sequence Metadata for Peptide Level Graphing (Step 10)
SKIP Using this Function. It is used in merge_metadata_pep function.}
\usage{
grab_seq_metadata_pep(df_in)
}
\arguments{
\item{df_in}{data frame with original data to match proteins and peptide
locations}
}
\value{
a data frame containing the MasterProteinAccessions, Sequence,
peptide, and start columns. ONLY used inside of merge_metadata_pep function.
}
\description{
Grab Sequence Metadata for Peptide Level Graphing (Step 10)
SKIP Using this Function. It is used in merge_metadata_pep function.
}
\examples{
new_df <- grab_seq_metadata_pep(raw_data)
}
