% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_seq_metadata_pep.R
\name{grab_seq_metadata_pep}
\alias{grab_seq_metadata_pep}
\title{grab_seq_metadata_pep}
\usage{
grab_seq_metadata_pep(df_in)
}
\arguments{
\item{df_in}{data frame with original data to match proteins and peptide
locations}
}
\value{
a data frame containing the MasterProteinAccessions, Sequence,
peptide, and start columns
}
\description{
grab_seq_metadata_pep
}
\examples{
new_df <- grab_seq_metadata_pep(raw_data)
}
