% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_data_res.R
\name{graphing_data_res}
\alias{graphing_data_res}
\title{Filter Data for Graphing the Extent of Modification at the Residue Level (Step 14)}
\usage{
graphing_data_res(df_in, pd_data_fasta_merged)
}
\arguments{
\item{df_in}{Data frame with residue level EOM calculations and raw data that
was merged with the FASTA file to filter for quantifiable modifications}

\item{pd_data_fasta_mergedraw}{Raw data merged with FASTA file}
}
\value{
A data frame containing quantifiable residue level modifications
}
\description{
Filter Data for Graphing the Extent of Modification at the Residue Level (Step 14)
}
\examples{
data_graphing<- graphing_data_res(filtered_graphing_df_res)
}
