% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_data_res.R
\name{graphing_data_res}
\alias{graphing_data_res}
\title{graphing_data_res}
\usage{
graphing_data_res(df_in)
}
\arguments{
\item{df_in}{data frame with residue level EOM calculations to filter
for quantifiable modifications}
}
\value{
a data frame containing quantifiable resiude level modifications
}
\description{
graphing_data_res
}
\examples{
data_graphing<- graphing_data_res(filtered_graphing_df_res)
}
