\name{output_folder}
\alias{output_folder}
\title{Select a folder for your tables and figures to be saved}
\description{
  This function prompts the user to select a folder on their computer using a dialog box.
}
\usage{
  output_folder()
}
\value{
  Returns the path of the selected folder as a character string.
}
\examples{
  # Call the function to select a folder
  selected_path <- output_folder()

}
