% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fasta.R
\name{parse_fasta}
\alias{parse_fasta}
\title{parse_fasta}
\usage{
parse_fasta(fasta_in)
}
\arguments{
\item{fasta_in}{FASTA file in the form of a dataframe}
}
\value{
a condensed dataframe with the protein sequence and the uniprot ID
}
\description{
parse_fasta
}
\examples{
filtered_FASTA <- parse_fasta(FASTA)
}
