% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graphs_pep.R
\name{save_graphs_pep}
\alias{save_graphs_pep}
\title{save_graphs_pep}
\usage{
save_graphs_pep(df, file_output, excel_filename)
}
\arguments{
\item{file_output}{The file path to save the graphs}

\item{excel_filename}{The file path to save the excel file}

\item{df_in_pep}{Bar graphs corresponding to peptides that had quantifiable
extent of modification}
}
\value{
A series of bar graphs corresponding to peptides that had
quantifiable extent of modification
}
\description{
save_graphs_pep
}
\examples{
save_graphs_pep(df_in_pep, file_output, excel_filename)
}
