% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_graphs_res.R
\name{save_graphs_res}
\alias{save_graphs_res}
\title{save_graphs_res}
\usage{
save_graphs_res(df, file_output, excel_filename)
}
\arguments{
\item{df}{A data frame that contains the data to be plotted}

\item{file_output}{The file path to save the graphs}

\item{excel_filename}{The file name to save the graphs}
}
\value{
A series of bar graphs that correspond to the residues with
quantifiable modifications
}
\description{
save_graphs_res
}
\examples{
save_graphs_res(df_in_pep, file_output, excel_filename)
}
