% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_spectrum_filesTMT.R
\name{split_spectrum_filesTMT}
\alias{split_spectrum_filesTMT}
\title{Extract Spectrum File Identifiers for TMT Data}
\usage{
split_spectrum_filesTMT(df_in)
}
\arguments{
\item{df_in}{}
}
\value{
A data frame containing the split spectrum files.
}
\description{
Extract Spectrum File Identifiers for TMT Data
}
\examples{
mod_data_fasta_merged <- split_spectrum_filesTMT(mod_data_fasta_merged)
}
